Monitoring_Board/requirements.txt
gradio
matplotlib
requests
vaderSentiment
pillow
tradingeconomics
